Issue in datatypes for API ETLJobs resolved

+++ PEI_ETL.Core/Entities/ETLJobs.cs
@@ -10,9 +10,9 @@ namespace PEI_ETL.Core.Entities
         public string Batch_Name { get; set; }
         public string Batch_Step { get; set; }
 
-        public int Job_Step_Seq_No { get; set; }
+        public decimal Job_Step_Seq_No { get; set; }
 
-        public int Job_Stage { get; set; }
+        public decimal Job_Stage { get; set; }
 
         public char Is_Step_Active { get; set; }
     }
